+++ src/bin/osm-lump-ways/cli_args.rs
@@ -5,6 +5,7 @@ use std::path::PathBuf;
 use osm_lump_ways::dij::SplitPathsMethod;
 use osm_lump_ways::tagfilter;
 use osm_lump_ways::taggrouper;
+use osm_lump_ways::way_group::MinLengthFilter;
 
 /// Group OSM ways based on shared tags into GeoJSON MultiLineStrings
 ///
@@ -98,6 +99,13 @@ pub struct Args {
     /// Only include (in the output) lines which are this length (in metres) or more.
     #[arg(long, value_name = "NUMBER")]
     pub min_length_m: Option<f64>,
+
+    /// Exclude groups which have less than this length. Can accept numbers (metres), or suffixed
+    /// with `km`. Special suffx `%longest` is adaptive to the longest group there is.
+    /// `--min-length 50%longest` includes the longest group, and any group which is at least as
+    /// long as 50% of the length of the longest.
+    #[arg(long, conflicts_with = "min_length_m")]
+    pub min_length: Option<MinLengthFilter>,
 
     /// Only include (in the output) lines which are this length (in metres) or less.
     #[arg(long, value_name = "NUMBER")]